+++ pkgs/http2/lib/transport.dart
@@ -303,8 +303,9 @@ class TransportException implements Exception {
 class TransportConnectionException extends TransportException {
   final int errorCode;
 
-  TransportConnectionException(this.errorCode, String details)
-      : super('Connection error: $details');
+  TransportConnectionException(int errorCode, String details)
+      : errorCode = errorCode,
+        super('Connection error: $details (errorCode: $errorCode)');
 }
 
 /// An exception thrown when a HTTP/2 stream error occured.